refactor: use a test_utils for build and preview
Also introduces a `webpack_config` shared function that can be used
instead of a file to store the webpack configuration.
