--- tests/fixtures/webpack_multiple_config/webpack.config.js
@@ -1,4 +1,0 @@
-module.exports = [
-  { entry: "./a.js" },
-  { entry: "./b.js" }
-]
