--- tests/fixtures/webpack_multiple_specify_config/package.json
@@ -1,1 +1,0 @@
-{}
