--- tests/fixtures/webpack_multiple_specify_config/webpack.worker.js
@@ -1,4 +1,0 @@
-module.exports = [
-    { entry: "./a.js" },
-    { entry: "./b.js" }
-  ]
