+++ tests/fixtures/webpack_simple_js/package.json
@@ -1,1 +1,3 @@
-{}
+{
+  "main": "./index.js"
+}
