--- tests/fixtures/webpack_simple_js/webpack.config.js
@@ -1,1 +1,0 @@
-module.exports = { entry: "./index.js" };
